+++ packages/ng-diagram/projects/ng-diagram/src/core/src/input-events/handlers/box-selection/box-selection.event.ts
@@ -1,0 +1,6 @@
+import { BasePointerInputEvent, InputEventPhase } from '../../input-events.interface';
+
+export interface BoxSelectionEvent extends BasePointerInputEvent {
+  name: 'boxSelection';
+  phase: InputEventPhase;
+}