+++ packages/ng-diagram/projects/ng-diagram/src/core/src/input-events/handlers/index.ts
@@ -1,3 +1,4 @@
+export * from './box-selection';
 export * from './copy';
 export * from './cut';
 export * from './delete-selection';