+++ CHANGELOG.md
@@ -11,7 +11,10 @@ and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0
 
 - Zoom to fit feature with configurable padding and option to automatically apply on model initialization ([#386](https://github.com/synergycodes/ng-diagram/pull/386))
 - Environment layer for unified environment-related functionalities ([#350](https://github.com/synergycodes/ng-diagram/pull/350))
-- Helpers for node relationships and traversal ([#395]https://github.com/synergycodes/ng-diagram/pull/395)
+- Helpers for node relationships and traversal ([#395](https://github.com/synergycodes/ng-diagram/pull/395))
+- Box selection for selecting multiple nodes at once ([#374](https://github.com/synergycodes/ng-diagram/pull/374))
+- Implemented multiple event hooks for ng-diagram ([#387](https://github.com/synergycodes/ng-diagram/pull/387))
+- Configurable built-in grid background ([#397](https://github.com/synergycodes/ng-diagram/pull/397))
 
 ### Changed
 
@@ -19,6 +22,7 @@ and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0
 - Improved documentation examples structure for consistency ([#360](https://github.com/synergycodes/ng-diagram/pull/360))
 - Unified documentation styles ([#357](https://github.com/synergycodes/ng-diagram/pull/357))
 - Redirected documentation root to quick-start page and reordered Intro articles ([#370](https://github.com/synergycodes/ng-diagram/pull/370))
+- Changed default behavior for resizable and rotatable properties on diagram nodes ([#374](https://github.com/synergycodes/ng-diagram/pull/374))
 
 ### Fixed